requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.1
django-elasticsearch-dsl==8.0
django-haystack==3.3.0
elastic-transport==8.15.0
elasticsearch==8.15.1
elasticsearch-dsl==8.15.3
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
